requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
qdrant-client==1.7.0
docling==1.5.0
sentence-transformers==2.2.2
langchain-text-splitters==0.0.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.24.3
aiofiles==23.2.0
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10
langchain-openai==0.0.5
langchain-google-genai==0.0.6
langchain-qdrant==0.0.1
langchain-huggingface==0.0.5
